refactor(state): share reconciliation decisions
Extract completion/cancel precedence and its decision results from the reaper so managed-primary reconciliation no longer reaches through a private import.

Refs #322
